+++ src/extension.ts
@@ -1,25 +1,21 @@
 /*
 	Christian Larsen, 2025
+	"RPG structure"
+	extension.ts
 */
 
 import * as vscode from 'vscode';
-import { handleInsert, getInitPlaceholder, updateContext } from './extension-util';
+import { handleInsert, updateContext, deleteField, insertField, insertSubstructure, insertFieldSubstructure } from './extension-util';
 import { header, fields } from './rpg-structure-model';
 import { HeaderTreeDataProvider, StructureItem, FieldsTreeDataProvider, FieldItem } from './extension-providers';
 
 // Function that activates the extension
 export function activate(context: vscode.ExtensionContext) {
 
-	console.log("✅ RPG Structure extension activated");
-
 	updateContext(vscode.window.activeTextEditor);
 	vscode.window.onDidChangeActiveTextEditor(editor => {
 		updateContext(editor);
 	});
-
-	// Editor and position to insert the code
-	let editor = vscode.window.activeTextEditor;
-	const insertPosition = editor?.selection.active;
 
 	// "Header" provider
 	const provider = new HeaderTreeDataProvider();
@@ -31,12 +27,10 @@ export function activate(context: vscode.ExtensionContext) {
 
 	// Check if the editor has been changed, so maybe I have to remove the extension from the activity bar
 	vscode.window.onDidChangeActiveTextEditor(editor => {
-		const isRpgle = editor?.document.languageId === 'rpgle' || editor?.document.languageId === 'sqlrpgle';
-		// ???? vscode.commands.executeCommand('setContext', 'rpgstructure.showContainer', isRpgle);
 		editor = vscode.window.activeTextEditor;
 	});
 
-	// Subscriptions
+	// Subscriptions (commands)
 	context.subscriptions.push(
 
 		// Structure data =================================================================================
@@ -97,78 +91,20 @@ export function activate(context: vscode.ExtensionContext) {
 
 		// Add Field ======================================================================================
 		vscode.commands.registerCommand('rpgStructure.addField', async () => {
-			const name = await vscode.window.showInputBox({ prompt: 'Field name' });
-			if (!name) return;
-
-			const type = await vscode.window.showQuickPick(['char', 'varchar', 'zoned', 'dec', 'ind', 'packed'], {
-				placeHolder: 'Choose field type'
-			});
-			if (!type) return;
-
-			let length: string | undefined;
-			if (type === 'char' || type === 'varchar' || type === 'zoned' || type === 'dec' || type === 'packed') {
-
-				const lenStr = await vscode.window.showInputBox({
-					prompt: 'Field length',
-					placeHolder: type === 'zoned' || type === 'packed' ? 'e.g. 5 or 13:2' : 'e.g. 10',
-					validateInput: (input) => {
-						if (!input) return 'Length is required';
-
-						// Simple integer
-						if (/^\d+$/.test(input)) return undefined;
-
-						// Format N:N for zoned/packed only
-						if ((type === 'zoned' || type === 'packed') && /^\d+:\d+$/.test(input)) {
-							const [whole, decimal] = input.split(':').map(Number);
-							if (decimal >= whole) return 'Decimal part must be smaller than total length';
-							return undefined;
-						};
-
-						return 'Invalid format';
-					}
-				});
-				length = lenStr;
-			};
-
-			const init = await vscode.window.showInputBox({
-				prompt: 'Init value (optional)',
-				placeHolder: getInitPlaceholder(type),
-				validateInput: (input) => {
-					if (!input) return undefined; // opcional
-
-					if (type === 'char' || type === 'varchar') {
-						if (!/^'.*'$/.test(input)) return 'String must be enclosed in single quotes';
-						const content = input.slice(1, -1); // remove quotes
-						const maxLength = length ? parseInt(length.split(':')[0]) : 0;
-						if (content.length > maxLength) return `String too long (max ${maxLength} chars)`;
-						return undefined;
-					};
-					if (type === 'ind') {
-						const validValues = ['1', '0', '*on', '*off'];
-						if (!validValues.includes(input.toLowerCase()))
-							return "Must be '1', '0', '*on', or '*off'";
-						return undefined;
-					};
-					if (type === 'zoned' || type === 'packed') {
-						if (!/^\d+(\.\d+)?$/.test(input))
-							return 'Must be a numeric value (e.g., 42 or 13.5)';
-						return undefined;
-					};
-					return undefined;
-				}
-			});
-
-			const newField = new FieldItem(name, type, length, init);
-			fieldsProvider.addField(newField);
-
-			vscode.commands.executeCommand('setContext', 'rpgStructure.hasFields', fields.length > 0);
-
+			// Adds a field to the "fieldsProvider", in the last position (fields.length)
+			insertField(false, fields.length, fieldsProvider);
 		}),
 
-		// Clean Header ======================================================================================
+		// Add Substructure ===============================================================================
+		vscode.commands.registerCommand('rpgStructure.addSubstructure', async () => {
+			// Adds a field (marked as "structure") to the "fieldsProvider", always in the last position
+			insertSubstructure(fieldsProvider);
+		}),
+
+		// Clean Header ===================================================================================
 		vscode.commands.registerCommand('rpgStructure.cleanHeader', () => {
 
-			// Clean the header structure
+			// Cleans the header structure
 			header.name = '';
 			header.type = '';
 			header.dimension = 0;
@@ -176,12 +112,13 @@ export function activate(context: vscode.ExtensionContext) {
 			// Refresh the provider
 			provider.refresh();
 
+			// Sets "rpgStructure.hasHeader" to true or false
 			vscode.commands.executeCommand('setContext', 'rpgStructure.hasHeader',
 				(header.name !== '') && (header.dimension !== 0) && (header.type !== ''));
 
 		}),
 
-		// Clean Fields ======================================================================================
+		// Clean Fields ===================================================================================
 		vscode.commands.registerCommand('rpgStructure.cleanFields', () => {
 
 			// Clean the fields
@@ -197,6 +134,7 @@ export function activate(context: vscode.ExtensionContext) {
 		// Create Structure ===============================================================================
 		vscode.commands.registerCommand('rpgStructure.generate', () => {
 
+			// Retrieves the editor and the insert position to generate the code
 			const editor = vscode.window.activeTextEditor;
 			const insertPosition = editor?.selection.active;
 
@@ -204,6 +142,18 @@ export function activate(context: vscode.ExtensionContext) {
 			if (editor && insertPosition) {
 				handleInsert(editor, insertPosition, header.name, header.type, header.dimension.toString(), fields);
 			};
+		}),
+
+		// Adds Field to Substructure =====================================================================
+		vscode.commands.registerCommand('rpgStructure.addFieldToSubstructure', (item: FieldItem) => {
+			
+			insertFieldSubstructure(item.idNumber, fieldsProvider);
+		}),
+
+		// Delete Field ===================================================================================
+		vscode.commands.registerCommand('rpgStructure.deleteField', (item: FieldItem) => {
+			// Deletes the selected item from the "fieldsProvider"
+			deleteField(item, fieldsProvider);
 		})
 	);
 };